Use UnaryScalar to implement negate functions
