+++ sql/src/main/java/io/crate/expression/scalar/arithmetic/NegateFunctions.java
@@ -1,0 +1,94 @@
+/*
+ * Licensed to Crate under one or more contributor license agreements.
+ * See the NOTICE file distributed with this work for additional
+ * information regarding copyright ownership.  Crate licenses this file
+ * to you under the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.  You may
+ * obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied.  See the License for the specific language governing
+ * permissions and limitations under the License.
+ *
+ * However, if you have executed another commercial license agreement
+ * with Crate these terms will supersede the license and you may use the
+ * software solely pursuant to the terms of the relevant commercial
+ * agreement.
+ */
+
+package io.crate.expression.scalar.arithmetic;
+
+import io.crate.expression.scalar.ScalarFunctionModule;
+import io.crate.expression.scalar.UnaryScalar;
+import io.crate.expression.symbol.FuncArg;
+import io.crate.expression.symbol.Function;
+import io.crate.expression.symbol.Symbol;
+import io.crate.metadata.FunctionIdent;
+import io.crate.metadata.FunctionImplementation;
+import io.crate.metadata.FunctionInfo;
+import io.crate.metadata.FunctionResolver;
+import io.crate.metadata.functions.params.FuncParams;
+import io.crate.metadata.functions.params.Param;
+import io.crate.types.DataType;
+import io.crate.types.DoubleType;
+import io.crate.types.FloatType;
+import io.crate.types.IntegerType;
+import io.crate.types.LongType;
+import io.crate.types.ShortType;
+
+import javax.annotation.Nullable;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+
+public final class NegateFunctions {
+
+    private static final String NAME = "_negate";
+
+    private static final FunctionResolver RESOLVER = new FunctionResolver() {
+
+        private final FuncParams funcParams = FuncParams.builder(Param.NUMERIC).build();
+
+        @Override
+        public FunctionImplementation getForTypes(List<DataType> dataTypes) throws IllegalArgumentException {
+            DataType dataType = dataTypes.get(0);
+            switch (dataType.id()) {
+                case DoubleType.ID:
+                    return new UnaryScalar<Double, Double>(NAME, dataType, dataType, x -> x * -1);
+                case FloatType.ID:
+                    return new UnaryScalar<Float, Float>(NAME, dataType, dataType, x -> x * -1);
+                case ShortType.ID:
+                    return new UnaryScalar<Short, Short>(NAME, dataType, dataType, x -> (short) (x * -1));
+                case IntegerType.ID:
+                    return new UnaryScalar<Integer, Integer>(NAME, dataType, dataType, x -> x * -1);
+                case LongType.ID:
+                    return new UnaryScalar<Long, Long>(NAME, dataType, dataType, x -> x * -1);
+                default:
+                    throw new IllegalArgumentException("Cannot negate values of type " + dataType.getName());
+            }
+        }
+
+        @Nullable
+        @Override
+        public List<DataType> getSignature(List<? extends FuncArg> dataTypes) {
+            return funcParams.match(dataTypes);
+        }
+    };
+
+    public static void register(ScalarFunctionModule module) {
+        module.register(NAME, RESOLVER);
+    }
+
+    public static Function createFunction(Symbol argument) {
+        FunctionInfo info = new FunctionInfo(
+            new FunctionIdent(NAME, Collections.singletonList(argument.valueType())),
+            argument.valueType()
+        );
+        //noinspection ArraysAsListWithZeroOrOneArgument # must use mutable list for arguments
+        return new Function(info, Arrays.asList(argument));
+    }
+}
